requirements.txt
requests==2.28.1
flake8==5.0.4
beautifulsoup4==4.11.1
lxml==4.9.1
python-dotenv==0.20.0
python-telegram-bot==13.13
emoji==2.0.0
selenium=4.4.3